Re-ogranized sprint 1
moved sprint specific files into folder, updated test cases and test gedcom files.
