+++ Sprint1/US_10_test.ged
@@ -1,0 +1,135 @@
+0 NOTE My name is Yunfan Ye
+0 HEAD
+1 SOUR Family Echo
+2 WWW http://www.familyecho.com/
+1 FILE My Family
+1 DATE 9 SEP 2018
+1 DEST ANSTFILE
+1 GEDC
+2 VERS 5.5.1
+2 FORM LINEAGE-LINKED
+1 SUBM @I1@
+2 NAME Yunfan Ye
+1 SUBN
+1 CHAR UTF-8
+0 @I1@ INDI
+1 NAME Harry /Potter/
+2 GIVN Harry
+2 SURN Potter
+2 _MARNM Potter
+1 SEX M
+1 BIRT
+2 DATE 2 DEC 1969
+1 FAMS @F1@
+1 FAMS @F2@
+1 FAMC @F3@
+0 @I2@ INDI
+1 NAME Lily /Evans/
+2 GIVN Lily
+2 SURN Evans
+2 _MARNM Evans
+1 SEX F
+1 BIRT
+2 DATE 17 OCT 1928
+1 FAMS @F3@
+1 FAMS @F4@
+1 FAMC @F5@
+0 @I3@ INDI
+1 NAME James /Zhang/
+2 GIVN James
+2 SURN Zhang
+2 _MARNM Potter
+1 SEX M
+1 BIRT
+2 DATE 5 AUG 1928
+1 FAMS @F3@
+0 @I4@ INDI
+1 NAME Ginny /Weasley/
+2 GIVN Ginny
+2 SURN Weasley
+2 _MARNM Weasley
+1 SEX F
+1 BIRT
+2 DATE 14 DEC 1968
+1 FAMS @F1@
+0 @I5@ INDI
+1 NAME Qiu /Zhang/
+2 GIVN Qiu
+2 SURN Zhang
+2 _MARNM Zhang
+1 SEX F
+1 BIRT
+2 DATE 3 FEB 1968
+1 DEAT Y
+2 DATE 9 SEP 2008
+1 FAMS @F2@
+0 @I6@ INDI
+1 NAME James /Potter/
+2 GIVN James
+2 SURN Potter
+2 _MARNM Potter
+1 SEX M
+1 BIRT
+2 DATE 12 SEP 2016
+1 FAMC @F1@
+0 @I7@ INDI
+1 NAME Albus /Potter/
+2 GIVN Albus
+2 SURN Potter
+2 _MARNM Potter
+1 SEX M
+1 BIRT
+2 DATE 13 OCT 2015
+1 FAMC @F1@
+0 @I8@ INDI
+1 NAME Lily /Potter/
+2 GIVN Lily
+2 SURN Potter
+2 _MARNM Potter
+1 SEX F
+1 BIRT
+2 DATE 10 DEC 2014
+1 FAMC @F1@
+0 @I9@ INDI
+1 NAME Michael /Fang/
+2 GIVN Michael
+2 SURN Fang
+2 _MARNM Fang
+1 SEX M
+1 BIRT
+2 DATE 14 DEC 2013
+1 FAMC @F2@
+0 @F1@ FAM
+1 HUSB @I1@
+1 WIFE @I4@
+1 CHIL @I6@
+1 CHIL @I7@
+1 CHIL @I8@
+1 MARR
+2 DATE 5 APR 2012
+2 PLAC Hogwarts
+1 _CURRENT Y
+0 @F2@ FAM
+1 HUSB @I1@
+1 WIFE @I5@
+1 CHIL @I9@
+1 MARR
+2 DATE 3 FEB 1978
+1 _SEPR
+2 DATE 9 NOV 2008
+1 _CURRENT N
+0 @F3@ FAM
+1 HUSB @I3@
+1 WIFE @I2@
+1 CHIL @I1@
+1 MARR
+2 DATE 13 JUL 1948
+2 PLAC Hogwarts
+1 _CURRENT Y
+0 @F4@ FAM
+1 WIFE @I2@
+1 _CURRENT N
+0 @F5@ FAM
+1 CHIL @I2@
+1 _CURRENT N
+0 TRLR
